Formalized queue push and retrieve. Untested.
Specifically, how the data moves from the original request to storage in the queue, from the queue to the handler, and the several levels of structure that we impose on it. One of these levels, close to the physical storage format, is transparently versioned, so that any changes in the future can be managed.

We're ready to start dealing with consumption operations.