Refactored cleanup.
We now use a Queue.Queue to track requests to be cleaned-up. We were keeping this in the KV, but we kept having misses b
etween the two servers trying to do cleanup.

We might try to simply "pop" off the KV queue (when limits are added to the in-order queue functionality with *etcd*).